# clients/openapi.yaml
openapi: 3.1.0
info:
  title: mem-dog API
  version: 4.0.0
  description: >
    Private AI system for multi-channel data ingestion, AI enrichment,
    temporal knowledge graph, and multi-signal semantic search.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.example.com
    description: Production
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key (md_*), JWT, or global API key
...
paths:
...
  /api/v1/data/{data_id}/tags:
    get:
      tags: [Tags]
      summary: Get tags for data item
      operationId: getTags
      parameters:
        - name: data_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tags
    put:
      tags: [Tags]
      summary: Replace all tags
      operationId: updateTags
      parameters:
        - name: data_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tags:
                  type: array
                  items:
                    type: string
              required: [tags]
      responses:
        '200':
          description: Updated
  /api/v1/data/{data_id}/tags/add:
    post:
      tags: [Tags]
      summary: Add tags (merge)
      operationId: addTags
      parameters:
        - name: data_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tags:
                  type: array
                  items:
                    type: string
              required: [tags]
      responses:
        '200':
          description: Tags added
  /api/v1/data/{data_id}/tags/remove:
    post:
      tags: [Tags]
      summary: Remove tags
      operationId: removeTags
      parameters:
        - name: data_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tags:
                  type: array
                  items:
                    type: string
              required: [tags]
      responses:
        '200':
          description: Tags removed
...
  /api/v1/tags/search:
    get:
      tags: [Tags]
      summary: Search by tags
      operationId: searchByTags
      parameters:
        - name: tags
          in: query
          required: true
          schema:
            type: string
        - name: match_all
          in: query
          schema:
            type: boolean
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Matching items

  # ---------- Access Control ----------
...
  /api/v1/ai/skills:
    get:
      tags: [Skills]
      summary: List skills
      operationId: listSkills
      parameters:
        - name: data_id
          in: query
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Skills
    post:
      tags: [Skills]
      summary: Create skill
      operationId: createSkill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Created
  /api/v1/ai/skills/{skill_id}:
    get:
      tags: [Skills]
      summary: Get skill
      operationId: getSkill
      parameters:
        - name: skill_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Skill
    put:
      tags: [Skills]
      summary: Update skill
      operationId: updateSkill
      parameters:
        - name: skill_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated
    delete:
      tags: [Skills]
      summary: Delete skill
      operationId: deleteSkill
      parameters:
        - name: skill_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted

  # ---------- Analysis Templates ----------
...
  /api/v1/store:
    get:
      tags: [Store]
      summary: List keys
      operationId: listStoreKeys
      parameters:
        - name: prefix
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Keys
  /api/v1/store/{key}:
    get:
      tags: [Store]
      summary: Get value by key
      operationId: getStoreValue
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Value
    put:
      tags: [Store]
      summary: Set value
      operationId: setStoreValue
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Set
    delete:
      tags: [Store]
      summary: Delete key
      operationId: deleteStoreValue
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted

  # ---------- Ingest ----------
...